+++ lib/public_api.ts
@@ -1,3 +1,6 @@
 export * from './loadable';
 export * from './loadable-observable';
 export * from './loadable-subject';
+
+export * from './creation-functions';
+export * from './operators';