+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.1.0
+
+- Add pipe to LoadableObservable
+- Add ldSwitchMap, ldFilter and ldMap operators
+- Add ldCombineLatest creation function
+
 # 0.0.2
 
 - Fixed location of types file so that package actually works with TS.